perf(aggbuffer): remove progress() function in AggBuffer
When the aggregation is on, the execution time of each AM handler is on the order of 300us, which is
too much for a backoff in AggBuffer. This might cause one thread holding a certain shared resource
for too much time and thus damaging the performance of all the other threads.